@@ -21,7 +21,7 @@
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-        <java.version>11</java.version>
+        <java.version>17</java.version>
         <javamelody.version>1.91.0</javamelody.version>
         <testcontainers.version>1.15.3</testcontainers.version>
     </properties>
