@@ -26,21 +26,16 @@ import java.util.stream.Collectors;
 public class DomainEntityHelperServiceImpl implements DomainEntityHelperService {
     final Logger logger = LoggerFactory.getLogger(this.getClass());
     private final NodeRepository nodeRepository;
-    private final ResourceRepository resourceRepository;
     private final NodeConnectionRepository nodeConnectionRepository;
-    private final NodeResourceRepository nodeResourceRepository;
     private final ResourceTypeRepository resourceTypeRepository;
     private final CachedUrlUpdaterService cachedUrlUpdaterService;
     private final CustomFieldService customFieldService;
 
-    public DomainEntityHelperServiceImpl(NodeRepository nodeRepository, ResourceRepository resourceRepository,
-            NodeConnectionRepository nodeConnectionRepository, NodeResourceRepository nodeResourceRepository,
-            ResourceTypeRepository resourceTypeRepository, CachedUrlUpdaterService cachedUrlUpdaterService,
-            CustomFieldService customFieldService) {
+    public DomainEntityHelperServiceImpl(NodeRepository nodeRepository,
+            NodeConnectionRepository nodeConnectionRepository, ResourceTypeRepository resourceTypeRepository,
+            CachedUrlUpdaterService cachedUrlUpdaterService, CustomFieldService customFieldService) {
         this.nodeRepository = nodeRepository;
-        this.resourceRepository = resourceRepository;
         this.nodeConnectionRepository = nodeConnectionRepository;
-        this.nodeResourceRepository = nodeResourceRepository;
         this.resourceTypeRepository = resourceTypeRepository;
         this.cachedUrlUpdaterService = cachedUrlUpdaterService;
         this.customFieldService = customFieldService;
@@ -60,16 +55,14 @@ public class DomainEntityHelperServiceImpl implements DomainEntityHelperService 
         case "subject":
         case "topic":
         case "node":
-            return nodeRepository.findFirstByPublicIdIncludingCachedUrlsAndTranslations(publicId).orElse(null);
         case "resource":
-            return resourceRepository.findFirstByPublicIdIncludingCachedUrlsAndTranslations(publicId).orElse(null);
+            return nodeRepository.findFirstByPublicIdIncludingCachedUrlsAndTranslations(publicId).orElse(null);
         case "node-connection":
         case "subject-topic":
         case "topic-subtopic":
-            return nodeConnectionRepository.findByPublicId(publicId);
         case "node-resource":
         case "topic-resource":
-            return nodeResourceRepository.findByPublicId(publicId);
+            return nodeConnectionRepository.findByPublicId(publicId);
         }
         throw new NotFoundServiceException("Entity of type not found");
     }
@@ -77,19 +70,12 @@ public class DomainEntityHelperServiceImpl implements DomainEntityHelperService 
     @Override
     public TaxonomyRepository getRepository(URI publicId) {
         switch (publicId.getSchemeSpecificPart().split(":")[0]) {
-        case "subject":
-        case "topic":
-        case "node":
+        case "subject", "topic", "node", "resource" -> {
             return nodeRepository;
-        case "resource":
-            return resourceRepository;
-        case "node-connection":
-        case "subject-topic":
-        case "topic-subtopic":
+        }
+        case "node-connection", "subject-topic", "topic-subtopic", "node-resource", "topic-resource" -> {
             return nodeConnectionRepository;
-        case "node-resource":
-        case "topic-resource":
-            return nodeResourceRepository;
+        }
         }
         throw new NotFoundServiceException(String.format("Unknown repository requested: %s", publicId));
     }
@@ -143,18 +129,12 @@ public class DomainEntityHelperServiceImpl implements DomainEntityHelperService 
             ((Node) domainEntity).getTranslations().forEach(NodeTranslation::getName);
             ((Node) domainEntity).getChildConnections();
             ((Node) domainEntity).getParentConnections();
-        } else if (domainEntity instanceof Resource) {
-            ((Resource) domainEntity).getTranslations().forEach(ResourceTranslation::getName);
-            ((Resource) domainEntity).getParentConnections();
-            ((Resource) domainEntity).getResourceResourceTypes();
-            ((Resource) domainEntity).getResourceResourceTypes().forEach(ResourceResourceType::getResourceType);
-            ((Resource) domainEntity).getResourceTypes().forEach(this::initializeFields);
+            ((Node) domainEntity).getResourceResourceTypes();
+            ((Node) domainEntity).getResourceResourceTypes().forEach(ResourceResourceType::getResourceType);
+            ((Node) domainEntity).getResourceTypes().forEach(this::initializeFields);
         } else if (domainEntity instanceof NodeConnection) {
             ((NodeConnection) domainEntity).getParent().ifPresent(this::initializeFields);
             ((NodeConnection) domainEntity).getChild().ifPresent(this::initializeFields);
-        } else if (domainEntity instanceof NodeResource) {
-            ((NodeResource) domainEntity).getNode().ifPresent(this::initializeFields);
-            ((NodeResource) domainEntity).getResource().ifPresent(this::initializeFields);
         } else if (domainEntity instanceof ResourceType) {
             ((ResourceType) domainEntity).getParent().ifPresent(this::initializeFields);
             ((ResourceType) domainEntity).getTranslations().forEach(ResourceTypeTranslation::getName);
@@ -179,14 +159,8 @@ public class DomainEntityHelperServiceImpl implements DomainEntityHelperService 
         if (domainEntity instanceof Node) {
             return updateNode((Node) domainEntity, cleanUp);
         }
-        if (domainEntity instanceof Resource) {
-            return updateResource((Resource) domainEntity, cleanUp);
-        }
         if (domainEntity instanceof NodeConnection) {
             return updateNodeConnection((NodeConnection) domainEntity, cleanUp);
-        }
-        if (domainEntity instanceof NodeResource) {
-            return updateNodeResource((NodeResource) domainEntity, cleanUp);
         }
         throw new IllegalArgumentException("Wrong type of element to update: " + domainEntity.getEntityName());
 
@@ -197,6 +171,7 @@ public class DomainEntityHelperServiceImpl implements DomainEntityHelperService 
         Node result;
         TaxonomyRepository<DomainEntity> repository = getRepository(node.getPublicId());
         Node existing = (Node) getEntityByPublicId(node.getPublicId());
+        node.getResourceTypes().forEach(this::ensureResourceTypesExists);
         if (existing == null) {
             mergeMetadata(null, node.getMetadata(), node.getPublicId(), cleanUp);
             result = repository.save(new Node(node));
@@ -209,6 +184,42 @@ public class DomainEntityHelperServiceImpl implements DomainEntityHelperService 
             existing.setNodeType(node.getNodeType());
             existing.setRoot(node.isRoot());
             existing.setContext(node.isContext());
+
+            // ResourceTypes
+            Collection<URI> typesToSet = new HashSet<>();
+            Collection<URI> typesToKeep = new HashSet<>();
+            List<URI> existingTypes = existing.getResourceTypes().stream().map(DomainEntity::getPublicId)
+                    .collect(Collectors.toList());
+            node.getResourceTypes().forEach(resourceType -> {
+                if (existingTypes.contains(resourceType.getPublicId())) {
+                    typesToKeep.add(resourceType.getPublicId());
+                } else {
+                    typesToSet.add(resourceType.getPublicId());
+                }
+            });
+            if (!typesToSet.isEmpty()) {
+                Map<URI, URI> reusedUris = node.getResourceResourceTypes().stream()
+                        .filter(resourceResourceType -> typesToSet
+                                .contains(resourceResourceType.getResourceType().getPublicId()))
+                        .collect(Collectors.toMap(
+                                resourceResourceType -> resourceResourceType.getResourceType().getPublicId(),
+                                ResourceResourceType::getPublicId));
+
+                Collection<ResourceResourceType> toRemove = new HashSet<>();
+                existing.getResourceResourceTypes().forEach(resourceResourceType -> {
+                    if (!typesToKeep.contains(resourceResourceType.getResourceType().getPublicId())) {
+                        toRemove.add(resourceResourceType);
+                    }
+                });
+                toRemove.forEach(existing::removeResourceResourceType);
+                typesToSet.forEach(uri -> {
+                    ResourceType resourceType = resourceTypeRepository.findByPublicId(uri);
+                    ResourceResourceType resourceResourceType = existing.addResourceType(resourceType);
+                    if (reusedUris.containsKey(resourceType.getPublicId())) {
+                        resourceResourceType.setPublicId(reusedUris.get(resourceType.getPublicId()));
+                    }
+                });
+            }
 
             // Translations
             Set<NodeTranslation> translations = new HashSet<>();
@@ -244,97 +255,6 @@ public class DomainEntityHelperServiceImpl implements DomainEntityHelperService 
                     deleteEntityByPublicId(nodeConnection.getPublicId());
                 }
             });
-            List<URI> resources = node.getNodeResources().stream().map(DomainEntity::getPublicId)
-                    .collect(Collectors.toList());
-            result.getNodeResources().forEach(nodeResource -> {
-                if (!resources.contains(nodeResource.getPublicId())) {
-                    deleteEntityByPublicId(nodeResource.getPublicId());
-                }
-            });
-        }
-        if (cleanUp) {
-            buildPathsForEntity(result.getPublicId());
-        }
-        return Optional.of(result);
-    }
-
-    @Transactional(propagation = Propagation.MANDATORY)
-    private Optional<DomainEntity> updateResource(Resource resource, boolean cleanUp) {
-        Resource result;
-        TaxonomyRepository<DomainEntity> repository = getRepository(resource.getPublicId());
-        Resource existing = (Resource) getEntityByPublicId(resource.getPublicId());
-        resource.getResourceTypes().forEach(this::ensureResourceTypesExists);
-        if (existing == null) {
-            mergeMetadata(null, resource.getMetadata(), resource.getPublicId(), cleanUp);
-            result = repository.save(new Resource(resource, true));
-        } else if (existing.equals(resource)) {
-            logger.debug("Resource " + resource.getPublicId() + " is equal, continue");
-            result = existing;
-        } else {
-            logger.debug("Updating resource " + resource.getPublicId());
-            existing.setName(resource.getName());
-            existing.setContentUri(resource.getContentUri());
-
-            // ResourceTypes
-            Collection<URI> typesToSet = new HashSet<>();
-            Collection<URI> typesToKeep = new HashSet<>();
-            List<URI> existingTypes = existing.getResourceTypes().stream().map(DomainEntity::getPublicId)
-                    .collect(Collectors.toList());
-            resource.getResourceTypes().forEach(resourceType -> {
-                if (existingTypes.contains(resourceType.getPublicId())) {
-                    typesToKeep.add(resourceType.getPublicId());
-                } else {
-                    typesToSet.add(resourceType.getPublicId());
-                }
-            });
-            if (!typesToSet.isEmpty()) {
-                Map<URI, URI> reusedUris = resource.getResourceResourceTypes().stream()
-                        .filter(resourceResourceType -> typesToSet
-                                .contains(resourceResourceType.getResourceType().getPublicId()))
-                        .collect(Collectors.toMap(
-                                resourceResourceType -> resourceResourceType.getResourceType().getPublicId(),
-                                ResourceResourceType::getPublicId));
-
-                Collection<ResourceResourceType> toRemove = new HashSet<>();
-                existing.getResourceResourceTypes().forEach(resourceResourceType -> {
-                    if (!typesToKeep.contains(resourceResourceType.getResourceType().getPublicId())) {
-                        toRemove.add(resourceResourceType);
-                    }
-                });
-                toRemove.forEach(existing::removeResourceResourceType);
-                typesToSet.forEach(uri -> {
-                    ResourceType resourceType = resourceTypeRepository.findByPublicId(uri);
-                    ResourceResourceType resourceResourceType = existing.addResourceType(resourceType);
-                    if (reusedUris.containsKey(resourceType.getPublicId())) {
-                        resourceResourceType.setPublicId(reusedUris.get(resourceType.getPublicId()));
-                    }
-                });
-            }
-
-            // Translations
-            Set<ResourceTranslation> translations = new HashSet<>();
-            for (ResourceTranslation resourceTranslation : resource.getTranslations()) {
-                Optional<ResourceTranslation> existingTranslation = existing.getTranslations().stream().filter(
-                        translation -> translation.getLanguageCode().equals(resourceTranslation.getLanguageCode()))
-                        .findFirst();
-                if (existingTranslation.isPresent()) {
-                    ResourceTranslation rt = existingTranslation.get();
-                    rt.setName(resourceTranslation.getName());
-                    translations.add(rt);
-                } else {
-                    translations.add(new ResourceTranslation(resourceTranslation, existing));
-                }
-            }
-            if (!translations.isEmpty()) {
-                existing.clearTranslations();
-                for (ResourceTranslation translation : translations) {
-                    existing.addTranslation(translation);
-                }
-            }
-
-            // Metadata
-            mergeMetadata(existing, resource.getMetadata(), resource.getPublicId(), cleanUp);
-            result = repository.save(existing);
         }
         if (cleanUp) {
             buildPathsForEntity(result.getPublicId());
@@ -379,46 +299,6 @@ public class DomainEntityHelperServiceImpl implements DomainEntityHelperService 
         }
         if (cleanUp) {
             buildPathsForEntity(result.getChild().get().getPublicId());
-        }
-        return Optional.of(result);
-    }
-
-    @Transactional(propagation = Propagation.MANDATORY)
-    private Optional<DomainEntity> updateNodeResource(NodeResource nodeResource, boolean cleanUp) {
-        NodeResource result;
-        TaxonomyRepository<DomainEntity> repository = getRepository(nodeResource.getPublicId());
-        TaxonomyRepository<DomainEntity> nodeRepository = getRepository(URI.create("urn:node:dummy"));
-        TaxonomyRepository<DomainEntity> resourceRepository = getRepository(URI.create("urn:resource:dummy"));
-
-        NodeResource existing = (NodeResource) getEntityByPublicId(nodeResource.getPublicId());
-        if (existing == null) {
-            mergeMetadata(null, nodeResource.getMetadata(), nodeResource.getPublicId(), cleanUp);
-            // Use correct objects when copying
-            nodeResource.setNode((Node) nodeRepository.findByPublicId(nodeResource.getNode().get().getPublicId()));
-            nodeResource.setResource(
-                    (Resource) resourceRepository.findByPublicId(nodeResource.getResource().get().getPublicId()));
-            NodeResource connection = new NodeResource(nodeResource);
-            result = repository.save(connection);
-        } else {
-            /*
-             * if (existing.equals(nodeResource)) { return Optional.of(existing); }
-             */
-            logger.debug("Updating noderesource " + nodeResource.getPublicId());
-            existing.setNode((Node) nodeRepository.findByPublicId(nodeResource.getNode().get().getPublicId()));
-            existing.setResource(
-                    (Resource) resourceRepository.findByPublicId(nodeResource.getResource().get().getPublicId()));
-            existing.setRank(nodeResource.getRank());
-            if (nodeResource.getRelevance().isPresent()) {
-                existing.setRelevance(nodeResource.getRelevance().get());
-            }
-            if (nodeResource.isPrimary().isPresent()) {
-                existing.setPrimary(nodeResource.isPrimary().get());
-            }
-            mergeMetadata(existing, nodeResource.getMetadata(), nodeResource.getPublicId(), cleanUp);
-            result = repository.save(existing);
-        }
-        if (cleanUp) {
-            buildPathsForEntity(result.getResource().get().getPublicId());
         }
         return Optional.of(result);
     }
