@@ -8,7 +8,7 @@
 package no.ndla.taxonomy.domain;
 
 public enum NodeType {
-    NODE("node"), SUBJECT("subject"), TOPIC("topic");
+    NODE("node"), SUBJECT("subject"), TOPIC("topic"), RESOURCE("resource");
 
     private String name;
 
