@@ -10,7 +10,7 @@ package no.ndla.taxonomy.domain;
 import java.util.Optional;
 
 public interface SortableResourceConnection<T extends DomainEntity> {
-    Optional<Resource> getResource();
+    Optional<Node> getResource();
 
     Optional<T> getParent();
 
