@@ -8,7 +8,7 @@
 package no.ndla.taxonomy.repositories;
 
 import no.ndla.taxonomy.domain.NodeConnection;
-import no.ndla.taxonomy.domain.Resource;
+import no.ndla.taxonomy.domain.NodeType;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.Query;
@@ -20,25 +20,73 @@ import java.util.Optional;
 import java.util.Set;
 
 public interface NodeConnectionRepository extends TaxonomyRepository<NodeConnection> {
-    @SuppressWarnings("SpringDataRepositoryMethodReturnTypeInspection")
-    @Query("SELECT nc.parent.id AS parentId, nc.child.id AS childId, nc.rank AS rank FROM NodeConnection nc"
-            + " JOIN nc.parent JOIN nc.child WHERE nc.parent.id IN :nodeId")
-    List<NodeTreeElement> findAllByNodeIdInIncludingTopicAndSubtopic(Set<Integer> nodeId);
+    @Query("""
+            SELECT DISTINCT nc
+            FROM NodeConnection nc
+            LEFT JOIN FETCH nc.metadata ncm LEFT JOIN FETCH ncm.grepCodes LEFT JOIN FETCH ncm.customFieldValues nccfv LEFT JOIN FETCH nccfv.customField
+            JOIN FETCH nc.parent p
+            LEFT JOIN FETCH p.metadata pm LEFT JOIN FETCH pm.grepCodes LEFT JOIN FETCH pm.customFieldValues pcfv LEFT JOIN FETCH pcfv.customField
+            JOIN FETCH nc.child c
+            LEFT JOIN FETCH c.metadata cm LEFT JOIN FETCH cm.grepCodes LEFT JOIN FETCH cm.customFieldValues ccfv LEFT JOIN FETCH ccfv.customField
+            LEFT JOIN FETCH nc.relevance rel
+            WHERE nc.parent.id IN :nodeId
+            AND ((:nodeTypes) IS NULL OR c.nodeType in :nodeTypes)
+            """)
+    List<NodeConnection> findAllByNodeIdInIncludingTopicAndSubtopic(Set<Integer> nodeId, List<NodeType> nodeTypes);
 
-    interface NodeTreeElement {
-        Integer getParentId();
+    @Query("""
+            SELECT DISTINCT nc FROM NodeConnection nc
+            LEFT JOIN FETCH nc.metadata ncm LEFT JOIN FETCH ncm.grepCodes LEFT JOIN FETCH ncm.customFieldValues nccfv LEFT JOIN FETCH nccfv.customField
+            JOIN FETCH nc.child r
+            LEFT JOIN FETCH r.metadata rm LEFT JOIN FETCH rm.grepCodes LEFT JOIN FETCH rm.customFieldValues rcfv LEFT JOIN FETCH rcfv.customField
+            LEFT JOIN FETCH nc.parent n
+            LEFT JOIN FETCH n.metadata nm LEFT JOIN FETCH nm.grepCodes LEFT JOIN FETCH nm.customFieldValues ncfv LEFT JOIN FETCH ncfv.customField
+            LEFT JOIN nc.relevance rel
+            LEFT JOIN r.resourceResourceTypes rrt
+            LEFT JOIN rrt.resourceType rt
+            LEFT JOIN r.cachedPaths
+            LEFT JOIN FETCH r.resourceResourceTypes rrtFetch
+            LEFT JOIN FETCH nc.relevance
+            LEFT JOIN FETCH rrtFetch.resourceType rtFetch
+            LEFT JOIN FETCH rtFetch.resourceTypeTranslations
+            WHERE n.id IN :nodeIds
+            AND (rt.publicId IN :resourceTypePublicIds)
+            AND r.nodeType = 'RESOURCE'
+            AND (:relevancePublicId IS NULL OR rel.publicId = :relevancePublicId)
+            """)
+    List<NodeConnection> getResourceBy(Set<Integer> nodeIds, Set<URI> resourceTypePublicIds, URI relevancePublicId);
 
-        Integer getChildId();
-
-        Integer getRank();
-    }
+    @Query("SELECT DISTINCT nc FROM NodeConnection nc " + NODE_CONNECTION_METADATA + " LEFT JOIN FETCH nc.parent n "
+            + NODE_METADATA + " LEFT JOIN FETCH nc.child r" + RESOURCE_METADATA
+            + " LEFT JOIN r.cachedPaths LEFT JOIN FETCH r.resourceResourceTypes rrtFetch"
+            + " LEFT JOIN FETCH rrtFetch.resourceType rtFetch LEFT JOIN FETCH rtFetch.resourceTypeTranslations"
+            + " WHERE n.id IN :nodeIds AND r.nodeType = 'RESOURCE'")
+    List<NodeConnection> getByResourceIds(Collection<Integer> nodeIds);
 
     @Query("SELECT nc FROM NodeConnection nc JOIN FETCH nc.parent JOIN FETCH nc.child JOIN FETCH nc.metadata m"
             + " LEFT JOIN m.grepCodes LEFT JOIN FETCH m.customFieldValues cvf LEFT JOIN cvf.customField")
     List<NodeConnection> findAllIncludingParentAndChild();
 
+    @Query("SELECT nc FROM NodeConnection nc JOIN FETCH nc.parent JOIN FETCH nc.child c JOIN FETCH nc.metadata m"
+            + " LEFT JOIN m.grepCodes LEFT JOIN FETCH m.customFieldValues cvf LEFT JOIN cvf.customField WHERE c.nodeType = :childNodeType")
+    List<NodeConnection> findAllByChildNodeType(NodeType childNodeType);
+
     @Query(value = "SELECT nc.id FROM NodeConnection nc ORDER BY nc.id", countQuery = "SELECT count(*) from NodeConnection")
     Page<Integer> findIdsPaginated(Pageable pageable);
+
+    @Query(value = """
+            SELECT nc
+            FROM NodeConnection nc
+            JOIN nc.child c
+            WHERE c.nodeType = :nodeType
+            ORDER BY nc.id
+            """, countQuery = """
+            SELECT count(nc)
+            FROM NodeConnection nc
+            JOIN nc.child c
+            WHERE c.nodeType = :nodeType
+            """)
+    Page<NodeConnection> findIdsPaginatedByChildNodeType(Pageable pageable, NodeType nodeType);
 
     @Query("SELECT DISTINCT nc FROM NodeConnection nc " + NODE_CONNECTION_METADATA + " JOIN FETCH nc.parent n "
             + NODE_METADATA + " JOIN FETCH nc.child c " + CHILD_METADATA + " WHERE nc.id in :ids")
