@@ -38,12 +38,11 @@ public class Nodes extends CrudControllerWithMetadata<Node> {
     private final NodeRepository nodeRepository;
     private final NodeConnectionRepository nodeConnectionRepository;
     private final NodeService nodeService;
-    private final ResourceService resourceService;
     private final RecursiveNodeTreeService recursiveNodeTreeService;
     private final TreeSorter treeSorter;
 
     public Nodes(NodeRepository nodeRepository, NodeConnectionRepository nodeConnectionRepository,
-            NodeService nodeService, CachedUrlUpdaterService cachedUrlUpdaterService, ResourceService resourceService,
+            NodeService nodeService, CachedUrlUpdaterService cachedUrlUpdaterService,
             RecursiveNodeTreeService recursiveNodeTreeService, TreeSorter treeSorter, MetadataService metadataService) {
         super(nodeRepository, cachedUrlUpdaterService, metadataService);
 
@@ -51,14 +50,13 @@ public class Nodes extends CrudControllerWithMetadata<Node> {
         this.nodeConnectionRepository = nodeConnectionRepository;
         this.nodeService = nodeService;
         this.recursiveNodeTreeService = recursiveNodeTreeService;
-        this.resourceService = resourceService;
         this.treeSorter = treeSorter;
     }
 
     @GetMapping
     @Operation(summary = "Gets all nodes")
     public List<EntityWithPathDTO> getAll(
-            @Parameter(description = "Filter by nodeType") @RequestParam(value = "nodeType", required = false) Optional<NodeType> nodeType,
+            @Parameter(description = "Filter by nodeType, could be a comma separated list :^)") @RequestParam(value = "nodeType", required = false) List<NodeType> nodeType,
             @Parameter(description = "ISO-639-1 language code", example = "nb") @RequestParam(value = "language", defaultValue = "", required = false) Optional<String> language,
             @Parameter(description = "Filter by contentUri") @RequestParam(value = "contentURI", required = false) Optional<URI> contentUri,
             @Parameter(description = "Only root level") @RequestParam(value = "isRoot", required = false) Optional<Boolean> isRoot,
@@ -200,7 +198,7 @@ public class Nodes extends CrudControllerWithMetadata<Node> {
             resourceTypeIdSet = new HashSet<>(Arrays.asList(resourceTypeIds));
         }
 
-        return resourceService.getResourcesByNodeId(nodeId, resourceTypeIdSet, relevance, language, recursive);
+        return nodeService.getResourcesByNodeId(nodeId, resourceTypeIdSet, relevance, language, recursive);
     }
 
     @PutMapping("/{id}/makeResourcesPrimary")
