@@ -33,16 +33,13 @@ import java.util.stream.Collectors;
 public class Topics extends CrudControllerWithMetadata<Node> {
     private final NodeRepository nodeRepository;
     private final NodeService nodeService;
-    private final ResourceService resourceService;
 
     public Topics(NodeRepository nodeRepository, NodeService nodeService,
-            CachedUrlUpdaterService cachedUrlUpdaterService, ResourceService resourceService,
-            MetadataService metadataService) {
+            CachedUrlUpdaterService cachedUrlUpdaterService, MetadataService metadataService) {
         super(nodeRepository, cachedUrlUpdaterService, metadataService);
 
         this.nodeRepository = nodeRepository;
         this.nodeService = nodeService;
-        this.resourceService = resourceService;
     }
 
     @GetMapping
@@ -55,8 +52,7 @@ public class Topics extends CrudControllerWithMetadata<Node> {
             @Parameter(description = "Filter by visible") @RequestParam(value = "isVisible", required = false) Optional<Boolean> isVisible) {
 
         MetadataFilters metadataFilters = new MetadataFilters(key, value, isVisible);
-        return nodeService.getNodes(language, Optional.of(NodeType.TOPIC), contentUri, Optional.empty(),
-                metadataFilters);
+        return nodeService.getNodes(language, List.of(NodeType.TOPIC), contentUri, Optional.empty(), metadataFilters);
     }
 
     @GetMapping("/search")
@@ -96,8 +92,7 @@ public class Topics extends CrudControllerWithMetadata<Node> {
     @Transactional
     public EntityWithPathDTO get(@PathVariable("id") URI id,
             @Parameter(description = "ISO-639-1 language code", example = "nb") @RequestParam(value = "language", required = false, defaultValue = "") String language) {
-        return new NodeDTO(nodeRepository.findFirstByPublicIdIncludingCachedUrlsAndTranslations(id)
-                .orElseThrow(() -> new NotFoundHttpResponseException("Topic was not found")), language);
+        return nodeService.getNode(id, language);
     }
 
     @PostMapping
@@ -181,7 +176,7 @@ public class Topics extends CrudControllerWithMetadata<Node> {
             resourceTypeIdSet = new HashSet<>(Arrays.asList(resourceTypeIds));
         }
 
-        return resourceService.getResourcesByNodeId(topicId, resourceTypeIdSet, relevance, language, recursive);
+        return nodeService.getResourcesByNodeId(topicId, resourceTypeIdSet, relevance, language, recursive);
     }
 
     @PutMapping("/{id}/makeResourcesPrimary")
