@@ -9,13 +9,13 @@ package no.ndla.taxonomy.rest.v1.commands;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.v3.oas.annotations.media.Schema;
-import no.ndla.taxonomy.domain.Resource;
+import no.ndla.taxonomy.domain.Node;
 import no.ndla.taxonomy.service.UpdatableDto;
 
 import java.net.URI;
 import java.util.Optional;
 
-public class ResourceCommand implements UpdatableDto<Resource> {
+public class ResourceCommand implements UpdatableDto<Node> {
     @JsonProperty
     @Schema(description = "If specified, set the id to this value. Must start with urn:resource: and be a valid URI. If omitted, an id will be assigned automatically.", example = "urn:resource:2")
     public URI id;
@@ -35,7 +35,7 @@ public class ResourceCommand implements UpdatableDto<Resource> {
     }
 
     @Override
-    public void apply(Resource entity) {
+    public void apply(Node entity) {
         if (getId().isPresent())
             entity.setPublicId(id);
         entity.setName(name);
