@@ -11,6 +11,7 @@ import no.ndla.taxonomy.domain.*;
 import no.ndla.taxonomy.repositories.ChangelogRepository;
 import no.ndla.taxonomy.repositories.NodeConnectionRepository;
 import no.ndla.taxonomy.repositories.NodeRepository;
+import no.ndla.taxonomy.rest.NotFoundHttpResponseException;
 import no.ndla.taxonomy.service.dtos.*;
 import no.ndla.taxonomy.service.exceptions.NotFoundServiceException;
 import no.ndla.taxonomy.service.task.Fetcher;
@@ -22,17 +23,16 @@ import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import javax.persistence.criteria.Join;
 import java.net.URI;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
+import java.util.Set;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
-
-import static org.springframework.data.jpa.domain.Specification.where;
 
 @Transactional(readOnly = true)
 @Service
@@ -45,6 +45,8 @@ public class NodeService implements SearchService<NodeDTO, Node, NodeRepository>
     private final TreeSorter topicTreeSorter;
     private final ChangelogRepository changelogRepository;
     private final DomainEntityHelperService domainEntityHelperService;
+    private final RecursiveNodeTreeService recursiveNodeTreeService;
+    private final TreeSorter treeSorter;
 
     private final ExecutorService executor = Executors.newSingleThreadExecutor();
 
@@ -53,9 +55,10 @@ public class NodeService implements SearchService<NodeDTO, Node, NodeRepository>
         executor.shutdown();
     }
 
-    public NodeService(NodeRepository nodeRepository, NodeConnectionRepository nodeConnectionRepository,
-            EntityConnectionService connectionService, VersionService versionService, TreeSorter topicTreeSorter,
-            ChangelogRepository changelogRepository, DomainEntityHelperService domainEntityHelperService) {
+    public NodeService(ChangelogRepository changelogRepository, DomainEntityHelperService domainEntityHelperService,
+            EntityConnectionService connectionService, NodeConnectionRepository nodeConnectionRepository,
+            NodeRepository nodeRepository, RecursiveNodeTreeService recursiveNodeTreeService,
+            TreeSorter topicTreeSorter, TreeSorter treeSorter, VersionService versionService) {
         this.nodeRepository = nodeRepository;
         this.nodeConnectionRepository = nodeConnectionRepository;
         this.connectionService = connectionService;
@@ -63,6 +66,8 @@ public class NodeService implements SearchService<NodeDTO, Node, NodeRepository>
         this.topicTreeSorter = topicTreeSorter;
         this.changelogRepository = changelogRepository;
         this.domainEntityHelperService = domainEntityHelperService;
+        this.recursiveNodeTreeService = recursiveNodeTreeService;
+        this.treeSorter = treeSorter;
     }
 
     @Transactional
@@ -76,66 +81,14 @@ public class NodeService implements SearchService<NodeDTO, Node, NodeRepository>
         nodeRepository.flush();
     }
 
-    public Specification<Node> base() {
-        return (root, query, criteriaBuilder) -> criteriaBuilder.isNotNull(root.get("id"));
-    }
-
-    public Specification<Node> nodeIsRoot() {
-        return (root, query, criteriaBuilder) -> criteriaBuilder.equal(root.get("root"), true);
-    }
-
-    public Specification<Node> nodeIsVisible(Boolean visible) {
-        return (root, query, criteriaBuilder) -> criteriaBuilder.equal(root.get("metadata").get("visible"), visible);
-    }
-
     public Specification<Node> nodeHasNodeType(NodeType nodeType) {
         return (root, query, criteriaBuilder) -> criteriaBuilder.equal(root.get("nodeType"), nodeType);
     }
 
-    public Specification<Node> nodeHasContentUri(URI contentUri) {
-        return (root, query, criteriaBuilder) -> criteriaBuilder.equal(root.get("contentUri"), contentUri);
-    }
-
-    public Specification<Node> nodeHasCustomKey(String key) {
-        return (root, query, criteriaBuilder) -> {
-            Join<Node, Metadata> nodeMetadataJoin = root.join("metadata");
-            Join<Metadata, CustomFieldValue> join = nodeMetadataJoin.join("customFieldValues");
-            return criteriaBuilder.equal(join.get("customField").get("key"), key);
-        };
-    }
-
-    public Specification<Node> nodeHasCustomValue(String value) {
-        return (root, query, criteriaBuilder) -> {
-            Join<Node, Metadata> nodeMetadataJoin = root.join("metadata");
-            Join<Metadata, CustomFieldValue> join = nodeMetadataJoin.join("customFieldValues");
-            return criteriaBuilder.equal(join.get("value"), value);
-        };
-    }
-
-    public List<EntityWithPathDTO> getNodes(Optional<String> language, Optional<NodeType> nodeType,
+    public List<EntityWithPathDTO> getNodes(Optional<String> language, List<NodeType> nodeType,
             Optional<URI> contentUri, Optional<Boolean> isRoot, MetadataFilters metadataFilters) {
-
-        final List<Node> filtered;
-        Specification<Node> specification = where(base());
-        if (isRoot.isPresent()) {
-            specification = specification.and(nodeIsRoot());
-        }
-        if (contentUri.isPresent()) {
-            specification = specification.and(nodeHasContentUri(contentUri.get()));
-        }
-        if (nodeType.isPresent()) {
-            specification = specification.and(nodeHasNodeType(nodeType.get()));
-        }
-        if (metadataFilters.getVisible().isPresent()) {
-            specification = specification.and(nodeIsVisible(metadataFilters.getVisible().get()));
-        }
-        if (metadataFilters.getKey().isPresent()) {
-            specification = specification.and(nodeHasCustomKey(metadataFilters.getKey().get()));
-        }
-        if (metadataFilters.getValue().isPresent()) {
-            specification = specification.and(nodeHasCustomValue(metadataFilters.getValue().get()));
-        }
-        filtered = nodeRepository.findAll(specification);
+        final List<Node> filtered = nodeRepository.findByNodeType(nodeType, metadataFilters.getVisible(),
+                metadataFilters.getKey(), metadataFilters.getValue(), contentUri, isRoot);
 
         return filtered.stream().distinct().map(n -> new NodeDTO(n, language.get())).collect(Collectors.toList());
     }
@@ -165,6 +118,85 @@ public class NodeService implements SearchService<NodeDTO, Node, NodeRepository>
         return topicTreeSorter.sortList(wrappedList);
     }
 
+    public NodeDTO getNode(URI publicId, String language) {
+        var node = getNode(publicId);
+        return new NodeDTO(node, language);
+    }
+
+    public Node getNode(URI publicId) {
+        return nodeRepository.findFirstByPublicIdIncludingCachedUrlsAndTranslations(publicId)
+                .orElseThrow(() -> new NotFoundHttpResponseException("Node was not found"));
+    }
+
+    public List<ResourceWithNodeConnectionDTO> getResourcesByNodeId(URI nodePublicId, Set<URI> resourceTypeIds,
+            URI relevancePublicId, String languageCode, boolean recursive) {
+        final var node = domainEntityHelperService.getNodeByPublicId(nodePublicId);
+
+        final Set<Integer> topicIdsToSearchFor;
+
+        // Add both topics and resourceTopics to a common list that will be sorted in a tree-structure based on rank at
+        // each level
+        final Set<ResourceTreeSortable<Node>> resourcesToSort = new HashSet<>();
+
+        // Populate a list of topic IDs we are going to fetch first, and then fetch the actual topics later
+        // This allows searching recursively without having to fetch the whole relation tree on each element in the
+        // recursive logic. It is also necessary to have the tree information later for ordering the result
+        if (recursive) {
+            final var nodeList = recursiveNodeTreeService.getRecursiveNodes(node);
+
+            nodeList.forEach(treeElement -> resourcesToSort.add(new ResourceTreeSortable<Node>("node", "node",
+                    treeElement.getId(), treeElement.getParentId().orElse(0), treeElement.getRank())));
+
+            topicIdsToSearchFor = nodeList.stream().map(RecursiveNodeTreeService.TreeElement::getId)
+                    .collect(Collectors.toSet());
+        } else {
+            topicIdsToSearchFor = Set.of(node.getId());
+        }
+
+        return filterNodeResourcesByIdsAndReturn(topicIdsToSearchFor, resourceTypeIds, relevancePublicId,
+                resourcesToSort, languageCode);
+    }
+
+    private List<ResourceWithNodeConnectionDTO> filterNodeResourcesByIdsAndReturn(Set<Integer> nodeIds,
+            Set<URI> resourceTypeIds, URI relevance, Set<ResourceTreeSortable<Node>> sortableListToAddTo,
+            String languageCode) {
+        final List<NodeConnection> nodeResources;
+
+        if (resourceTypeIds.size() > 0) {
+            nodeResources = nodeConnectionRepository.getResourceBy(nodeIds, resourceTypeIds, relevance);
+        } else {
+            var nodeResourcesStream = nodeConnectionRepository.getByResourceIds(nodeIds).stream();
+            if (relevance != null) {
+                final var isRequestingCore = "urn:relevance:core".equals(relevance.toString());
+                nodeResourcesStream = nodeResourcesStream.filter(nodeResource -> {
+                    final var resource = nodeResource.getChild().orElse(null);
+                    if (resource == null) {
+                        return false;
+                    }
+                    final var rel = nodeResource.getRelevance().orElse(null);
+                    if (rel != null) {
+                        return rel.getPublicId().equals(relevance);
+                    } else {
+                        return isRequestingCore;
+                    }
+                });
+            }
+            nodeResources = nodeResourcesStream.collect(Collectors.toList());
+        }
+
+        nodeResources.forEach(nodeResource -> sortableListToAddTo.add(new ResourceTreeSortable<Node>(nodeResource)));
+
+        // Sort the list, extract all the topicResource objects in between topics and return list of documents
+
+        return treeSorter.sortList(sortableListToAddTo).stream().map(ResourceTreeSortable::getResourceConnection)
+                .filter(Optional::isPresent).map(Optional::get)
+                .filter(connection -> ((NodeConnection) connection).getChild()
+                        .map(c -> c.getNodeType() == NodeType.RESOURCE).orElse(false))
+                .map(wrappedNodeResource -> new ResourceWithNodeConnectionDTO((NodeConnection) wrappedNodeResource,
+                        languageCode))
+                .collect(Collectors.toList());
+    }
+
     @Override
     public NodeRepository getRepository() {
         return nodeRepository;
@@ -186,11 +218,17 @@ public class NodeService implements SearchService<NodeDTO, Node, NodeRepository>
         final var node = nodeRepository.findFirstByPublicId(nodePublicId)
                 .orElseThrow(() -> new NotFoundServiceException("Node was not found"));
         if (recursive) {
-            node.getChildren().forEach(nc -> nc.getChild().map(n -> makeAllResourcesPrimary(n.getPublicId(), true)));
+            node.getChildren().forEach(nc -> {
+                nc.getChild().filter(n -> n.getNodeType() != NodeType.RESOURCE)
+                        .map(n -> makeAllResourcesPrimary(n.getPublicId(), true));
+            });
         }
-        node.getNodeResources().forEach(
-                nr -> connectionService.updateNodeResource(nr, nr.getRelevance().orElse(null), true, nr.getRank()));
-        return node.getNodeResources().stream()
+
+        node.getResourceChildren().forEach(cc -> {
+            connectionService.updateParentChild(cc, cc.getRelevance().orElse(null), cc.getRank(), Optional.of(true));
+        });
+
+        return node.getResourceChildren().stream()
                 .allMatch(resourceConnection -> resourceConnection.isPrimary().orElse(false));
     }
 
@@ -240,11 +278,6 @@ public class NodeService implements SearchService<NodeDTO, Node, NodeRepository>
             }
             changelogRepository.save(new Changelog(source, target, connection.getPublicId(), cleanUp));
         }
-        for (NodeResource nodeResource : node.getNodeResources()) {
-            nodeResource.getResource().map(
-                    child -> changelogRepository.save(new Changelog(source, target, child.getPublicId(), cleanUp)));
-            changelogRepository.save(new Changelog(source, target, nodeResource.getPublicId(), cleanUp));
-        }
         // When cleaning, node can be cleaned last to end with publish request to be stripped
         if (cleanUp) {
             changelogRepository.save(new Changelog(source, target, nodeId, true));
@@ -255,5 +288,12 @@ public class NodeService implements SearchService<NodeDTO, Node, NodeRepository>
         if (isRoot) {
             logger.info("Node " + nodeId + " added to changelog for publishing to " + target);
         }
+    }
+
+    public Node cloneNode(URI publicId, URI contentUri) {
+        final var node = getNode(publicId);
+        var cloned = new Node(node, false);
+        cloned.setContentUri(contentUri);
+        return nodeRepository.save(cloned);
     }
 }
