@@ -10,7 +10,7 @@ package no.ndla.taxonomy.service.dtos;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.v3.oas.annotations.media.Schema;
-import no.ndla.taxonomy.domain.NodeResource;
+import no.ndla.taxonomy.domain.NodeConnection;
 import no.ndla.taxonomy.domain.Relevance;
 import no.ndla.taxonomy.domain.exceptions.NotFoundException;
 
@@ -35,8 +35,8 @@ public class NodeWithResourceConnectionDTO extends NodeDTO {
     @JsonInclude(JsonInclude.Include.NON_NULL)
     private MetadataDto metadata;
 
-    public NodeWithResourceConnectionDTO(NodeResource nodeResource, String language) {
-        super(nodeResource.getNode().orElseThrow(() -> new NotFoundException("Node was not found")), language);
+    public NodeWithResourceConnectionDTO(NodeConnection nodeResource, String language) {
+        super(nodeResource.getParent().orElseThrow(() -> new NotFoundException("Node was not found")), language);
 
         this.isPrimary = nodeResource.isPrimary().orElse(false);
         this.connectionId = nodeResource.getPublicId();
