@@ -9,7 +9,7 @@ package no.ndla.taxonomy.service.dtos;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import no.ndla.taxonomy.domain.Node;
-import no.ndla.taxonomy.domain.NodeResource;
+import no.ndla.taxonomy.domain.NodeConnection;
 import no.ndla.taxonomy.domain.Relevance;
 import no.ndla.taxonomy.domain.exceptions.NotFoundException;
 
@@ -36,11 +36,11 @@ public class ResourceWithNodeConnectionDTO extends ResourceDTO {
 
     }
 
-    public ResourceWithNodeConnectionDTO(NodeResource nodeResource, String language) {
+    public ResourceWithNodeConnectionDTO(NodeConnection nodeResource, String language) {
         super(nodeResource.getResource().orElseThrow(() -> new NotFoundException("NodeResource was not found")),
                 language);
 
-        this.parentId = nodeResource.getNode().map(Node::getPublicId).orElse(null);
+        this.parentId = nodeResource.getParent().map(Node::getPublicId).orElse(null);
 
         this.connectionId = nodeResource.getPublicId();
         this.rank = nodeResource.getRank();
