@@ -9,7 +9,7 @@ package no.ndla.taxonomy.service.dtos;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.v3.oas.annotations.media.Schema;
-import no.ndla.taxonomy.domain.Resource;
+import no.ndla.taxonomy.domain.Node;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -27,10 +27,10 @@ public class ResourceWithParentsDTO extends ResourceDTO {
         super();
     }
 
-    public ResourceWithParentsDTO(Resource resource, String languageCode) {
+    public ResourceWithParentsDTO(Node resource, String languageCode) {
         super(resource, languageCode);
 
-        resource.getNodeResources().stream()
+        resource.getParentNodeConnections().stream()
                 .map(nodeResource -> new NodeWithResourceConnectionDTO(nodeResource, languageCode))
                 .forEach(parents::add);
     }
