@@ -9,7 +9,6 @@ package no.ndla.taxonomy.rest.v1;
 
 import no.ndla.taxonomy.domain.NodeType;
 import no.ndla.taxonomy.service.dtos.NodeDTO;
-import no.ndla.taxonomy.service.dtos.ResourceDTO;
 import org.junit.jupiter.api.Test;
 import org.springframework.mock.web.MockHttpServletResponse;
 
@@ -20,11 +19,11 @@ public class QueryTest extends RestTest {
 
     @Test
     public void can_get_resource_by_contentURI() throws Exception {
-        builder.resource(r -> r.publicId("urn:resource:1").contentUri("urn:article:345")
+        builder.node(NodeType.RESOURCE, r -> r.publicId("urn:resource:1").contentUri("urn:article:345")
                 .resourceType(rt -> rt.name("Subject material")));
 
         MockHttpServletResponse response = testUtils.getResource("/v1/queries/resources?contentURI=urn:article:345");
-        ResourceDTO[] resources = testUtils.getObject(ResourceDTO[].class, response);
+        var resources = testUtils.getObject(NodeDTO[].class, response);
 
         assertEquals(1, resources.length);
         assertEquals(resources[0].getResourceTypes().size(), 1);
@@ -33,20 +32,20 @@ public class QueryTest extends RestTest {
     @Test
     public void no_resources_matching_contentURI() throws Exception {
         MockHttpServletResponse response = testUtils.getResource("/v1/queries/resources?contentURI=urn:article:345");
-        ResourceDTO[] resources = testUtils.getObject(ResourceDTO[].class, response);
+        var resources = testUtils.getObject(NodeDTO[].class, response);
 
         assertEquals(0, resources.length);
     }
 
     @Test
     public void can_get_all_resources_matching_contentURI() throws Exception {
-        builder.resource(r -> r.publicId("urn:resource:1").contentUri("urn:article:345")
+        builder.node(NodeType.RESOURCE, r -> r.publicId("urn:resource:1").contentUri("urn:article:345")
                 .resourceType(rt -> rt.name("Subject material")));
 
-        builder.resource(r -> r.publicId("urn:resource:2").contentUri("urn:article:3"));
+        builder.node(NodeType.RESOURCE, r -> r.publicId("urn:resource:2").contentUri("urn:article:3"));
 
         MockHttpServletResponse response = testUtils.getResource("/v1/queries/resources?contentURI=urn:article:345");
-        ResourceDTO[] resources = testUtils.getObject(ResourceDTO[].class, response);
+        var resources = testUtils.getObject(NodeDTO[].class, response);
 
         assertEquals(1, resources.length);
         assertAnyTrue(resources, r -> "urn:resource:1".equals(r.getId().toString()));
@@ -54,11 +53,11 @@ public class QueryTest extends RestTest {
 
     @Test
     public void can_get_all_resource_types_for_a_resource() throws Exception {
-        builder.resource(r -> r.publicId("urn:resource:1").contentUri("urn:article:345")
+        builder.node(NodeType.RESOURCE, r -> r.publicId("urn:resource:1").contentUri("urn:article:345")
                 .resourceType(rt -> rt.name("Subject material")).resourceType(rt -> rt.name("Learning path")));
 
         MockHttpServletResponse response = testUtils.getResource("/v1/queries/resources?contentURI=urn:article:345");
-        ResourceDTO[] resources = testUtils.getObject(ResourceDTO[].class, response);
+        var resources = testUtils.getObject(NodeDTO[].class, response);
 
         assertEquals(1, resources.length);
         assertEquals(2, resources[0].getResourceTypes().size());
@@ -66,12 +65,13 @@ public class QueryTest extends RestTest {
 
     @Test
     public void can_get_translated_name_for_resource() throws Exception {
-        builder.resource(r -> r.publicId("urn:resource:1").name("Resource").translation("nb", tr -> tr.name("ressurs"))
-                .contentUri("urn:article:345").resourceType(rt -> rt.name("Subject material")));
+        builder.node(NodeType.RESOURCE,
+                r -> r.publicId("urn:resource:1").name("Resource").translation("nb", tr -> tr.name("ressurs"))
+                        .contentUri("urn:article:345").resourceType(rt -> rt.name("Subject material")));
 
         MockHttpServletResponse response = testUtils
                 .getResource("/v1/queries/resources?contentURI=urn:article:345&language=nb");
-        ResourceDTO[] resources = testUtils.getObject(ResourceDTO[].class, response);
+        var resources = testUtils.getObject(NodeDTO[].class, response);
 
         assertEquals(1, resources.length);
         assertEquals(resources[0].getResourceTypes().size(), 1);
@@ -80,13 +80,13 @@ public class QueryTest extends RestTest {
 
     @Test
     public void can_get_a_topic_matching_contentURI() throws Exception {
-        builder.resource(r -> r.publicId("urn:resource:1").contentUri("urn:article:345")
+        builder.node(NodeType.RESOURCE, r -> r.publicId("urn:resource:1").contentUri("urn:article:345")
                 .resourceType(rt -> rt.name("Subject material")));
 
         builder.node(NodeType.TOPIC, r -> r.publicId("urn:topic:2").contentUri("urn:article:345"));
 
         MockHttpServletResponse response = testUtils.getResource("/v1/queries/topics?contentURI=urn:article:345");
-        ResourceDTO[] resources = testUtils.getObject(ResourceDTO[].class, response);
+        var resources = testUtils.getObject(NodeDTO[].class, response);
 
         assertEquals(1, resources.length);
         assertAnyTrue(resources, r -> "urn:topic:2".equals(r.getId().toString()));
@@ -99,7 +99,7 @@ public class QueryTest extends RestTest {
         builder.node(NodeType.TOPIC, r -> r.publicId("urn:topic:3").contentUri("urn:article:345"));
 
         MockHttpServletResponse response = testUtils.getResource("/v1/queries/topics?contentURI=urn:article:345");
-        ResourceDTO[] resources = testUtils.getObject(ResourceDTO[].class, response);
+        var resources = testUtils.getObject(NodeDTO[].class, response);
 
         assertEquals(2, resources.length);
         assertAnyTrue(resources, r -> "urn:topic:2".equals(r.getId().toString()));
@@ -121,7 +121,7 @@ public class QueryTest extends RestTest {
 
         MockHttpServletResponse response = testUtils
                 .getResource("/v1/queries/topics?contentURI=urn:article:345&language=nb");
-        ResourceDTO[] resources = testUtils.getObject(ResourceDTO[].class, response);
+        var resources = testUtils.getObject(NodeDTO[].class, response);
 
         assertEquals(1, resources.length);
         assertEquals("Emne", resources[0].getName());
