@@ -7,7 +7,7 @@
 
 package no.ndla.taxonomy.rest.v1;
 
-import no.ndla.taxonomy.domain.Resource;
+import no.ndla.taxonomy.domain.NodeType;
 import no.ndla.taxonomy.service.dtos.ResourceDTO;
 import org.junit.jupiter.api.Test;
 import org.springframework.mock.web.MockHttpServletResponse;
@@ -23,8 +23,9 @@ public class ResourceTranslationsTest extends RestTest {
 
     @Test
     public void can_get_all_resources_with_translation() throws Exception {
-        builder.resource(r -> r.name("The inner planets").translation("nb", tr -> tr.name("De indre planetene")));
-        builder.resource(r -> r.name("Gas giants").translation("nb", tr -> tr.name("Gasskjemper")));
+        builder.node(NodeType.RESOURCE,
+                r -> r.name("The inner planets").translation("nb", tr -> tr.name("De indre planetene")));
+        builder.node(NodeType.RESOURCE, r -> r.name("Gas giants").translation("nb", tr -> tr.name("Gasskjemper")));
 
         MockHttpServletResponse response = testUtils.getResource("/v1/resources?language=nb");
         final var resources = testUtils.getObject(ResourceDTO[].class, response);
@@ -36,7 +37,7 @@ public class ResourceTranslationsTest extends RestTest {
 
     @Test
     public void can_get_single_resource_with_translation() throws Exception {
-        URI trigonometry = builder.resource(s -> s.name("introduction to trigonometry").translation("nb",
+        URI trigonometry = builder.node(NodeType.RESOURCE, s -> s.name("introduction to trigonometry").translation("nb",
                 tr -> tr.name("Introduksjon til trigonometri"))).getPublicId();
 
         MockHttpServletResponse response = testUtils.getResource("/v1/resources/" + trigonometry + "?language=nb");
@@ -47,14 +48,14 @@ public class ResourceTranslationsTest extends RestTest {
 
     @Test
     public void fallback_to_default_language() throws Exception {
-        URI id = builder.resource(t -> t.name("Introduction to algrebra")).getPublicId();
+        URI id = builder.node(NodeType.RESOURCE, t -> t.name("Introduction to algrebra")).getPublicId();
         final var resource = getResourceIndexDocument(id, "XX");
         assertEquals("Introduction to algrebra", resource.getName());
     }
 
     @Test
     public void can_get_default_language() throws Exception {
-        URI id = builder.resource(
+        URI id = builder.node(NodeType.RESOURCE,
                 t -> t.name("Introduction to algrebra").translation("nb", l -> l.name("Introduksjon til algebra")))
                 .getPublicId();
 
@@ -64,7 +65,7 @@ public class ResourceTranslationsTest extends RestTest {
 
     @Test
     public void can_add_translation() throws Exception {
-        Resource resource = builder.resource(t -> t.name("Introduction to algrebra"));
+        var resource = builder.node(NodeType.RESOURCE, t -> t.name("Introduction to algrebra"));
         URI id = resource.getPublicId();
 
         testUtils.updateResource("/v1/resources/" + id + "/translations/nb",
@@ -79,7 +80,7 @@ public class ResourceTranslationsTest extends RestTest {
 
     @Test
     public void can_delete_translation() throws Exception {
-        Resource resource = builder.resource(
+        var resource = builder.node(NodeType.RESOURCE,
                 t -> t.name("Introduction to algrebra").translation("nb", l -> l.name("Introduksjon til algebra")));
         URI id = resource.getPublicId();
 
@@ -90,7 +91,7 @@ public class ResourceTranslationsTest extends RestTest {
 
     @Test
     public void can_get_all_translations() throws Exception {
-        Resource resource = builder.resource(
+        var resource = builder.node(NodeType.RESOURCE,
                 t -> t.name("Introduction to algrebra").translation("nb", l -> l.name("Introduksjon til algebra"))
                         .translation("en", l -> l.name("Introduction to algrebra"))
                         .translation("de", l -> l.name("Introduktion bis Algebra")));
@@ -108,7 +109,7 @@ public class ResourceTranslationsTest extends RestTest {
 
     @Test
     public void can_get_single_translation() throws Exception {
-        Resource resource = builder.resource(
+        var resource = builder.node(NodeType.RESOURCE,
                 t -> t.name("Introduction to algrebra").translation("nb", l -> l.name("Introduksjon til algebra")));
         URI id = resource.getPublicId();
 
