@@ -9,7 +9,6 @@ package no.ndla.taxonomy.rest.v1;
 
 import no.ndla.taxonomy.domain.Node;
 import no.ndla.taxonomy.domain.NodeType;
-import no.ndla.taxonomy.domain.Resource;
 import no.ndla.taxonomy.domain.ResourceType;
 import no.ndla.taxonomy.domain.exceptions.IdFormatException;
 import no.ndla.taxonomy.service.URNValidator;
@@ -68,7 +67,7 @@ public class URNValidatorTest {
     @Test
     public void listingStyleURIIsAccepted() {
         URI id = URI.create("urn:resource:1:verktoy:blikkenslageren");
-        validator.validate(id, new Resource());
+        validator.validate(id, new Node(NodeType.RESOURCE));
     }
 
     @Test
